+++ components/footer.jsx
@@ -1,0 +1,6 @@
+export default function Footer() {
+    return (
+        <div className="footer">
+            <p>Jaskaran Sethy &copy; 2023</p>
+        </div>);
+}
