+++ components/header.jsx
@@ -1,0 +1,5 @@
+export default function Header() {
+    return (<div className="header">
+        <h1>Welcome to My Website</h1>
+    </div>);
+}
