+++ components/navbar.module.css
@@ -1,0 +1,3 @@
+.container {
+    margin: 0 auto 5rem;
+}
