+++ pages/blog/index.jsx
@@ -1,0 +1,11 @@
+import Layout, { Metadata, Header, NavBar, Content, Footer } from '../../components/layout';
+
+export default function HomePage() {
+    return (<Layout>
+            <Header />
+            <NavBar />
+            <Content />
+            <Footer />
+        </Layout>
+    );
+}
